Add attribute binding to int[] field
